fix: harden dashboard sandbox route in production
